refactor bit-node-modules path, add a new method getNodeModulesPathOfComponent() and remove duplication instances of this method
